refactor(server): #125 — extract MLS group handlers into routes/groups.rs
